refactor: CardPreviewSection 컴포넌트 삭제
