+++ src/feature/CardRegister/components/CardPreview/CardBrandLogo/CardBrandLogo.tsx
