+++ src/feature/CardRegister/components/CardPreview/CardExpiryDateDisplay/CardExpiryDateDisplay.stories.tsx
