+++ src/feature/CardRegister/components/CardPreview/CardExpiryDateDisplay/CardExpiryDateDisplay.tsx
