+++ src/feature/CardRegister/components/CardPreview/CardNumberDisplay/CardNumberDisplay.tsx
